Move 3 more files to correct architectural tiers (#603)
Continue pre-work for Epic #603 (Mechanical constraint enforcement).

Moves:
- waitForReactUpdates() → lib/async-utils.ts (generic async util → Tier 1)
  Breaks the interactive-engine → requirements-manager cycle.
- keyboard-shortcuts.hook.ts → components/docs-panel/ (single consumer → Tier 4)
- GuideResponseContext.tsx → docs-retrieval/ (domain-specific context → Tier 2)

All changes are mechanical: file moves and import path updates, no logic changes.
