+++ src/components/docs-panel/keyboard-shortcuts.hook.test.ts
@@ -1,6 +1,6 @@
 import { renderHook, cleanup, fireEvent } from '@testing-library/react';
 import { useKeyboardShortcuts } from './keyboard-shortcuts.hook';
-import type { LearningJourneyTab } from '../types/content-panel.types';
+import type { LearningJourneyTab } from '../../types/content-panel.types';
 
 describe('useKeyboardShortcuts', () => {
   const mockModel = {
