+++ src/interactive-engine/interactive.hook.ts
@@ -1,11 +1,7 @@
 import { useEffect, useCallback, useRef, useMemo } from 'react';
 import { addGlobalInteractiveStyles } from '../styles/interactive.styles';
-import {
-  waitForReactUpdates,
-  checkRequirements,
-  checkPostconditions,
-  RequirementsCheckOptions,
-} from '../requirements-manager';
+import { waitForReactUpdates } from '../lib/async-utils';
+import { checkRequirements, checkPostconditions, RequirementsCheckOptions } from '../requirements-manager';
 import { extractInteractiveDataFromElement } from '../lib/dom';
 import { InteractiveElementData } from '../types/interactive.types';
 import { INTERACTIVE_CONFIG } from '../constants/interactive-config';
